network: virtualnet: Add CreateRouterServerEnv()
The new added function sets up an environment with two Env objects: one
for router, which will be managed by shill; and one for server, which is
connected the router env. This setup is useful to verify the default
routes by checking the connectivity to services on the server env.

BUG=b:192436642
TEST=checked static IP addresses in the server Env are pingable on DUT

Change-Id: I5ee3dc88856daaebbb723f1033d73d7f47145595