ui: Put WebRTC Internals tab in separate window (in MeetCUJ)
This is a speculative fix for a test flake where the test fails to find
the download button on chrome://webrtc-internals. It may be related to
switching to the chrome://webrtc-internals tab when it is bogged down by
many meeting participants. This CL puts the chrome://webrtc-internals
tab in a separate window so that no such tab switching is necessary.

BUG=b:233674074
TEST=tast run <DUT> ui.MeetCUJ.*

Change-Id: If0c58f41f3c5e1d53201128de5bce3eb586556b0