drivefs: Use fixture for Chrome network feature
This adds a fixture that sets the necessary flags to enable Chrome
networking in DriveFS. I've also started using the fixture in the Google
Doc tests, to ensure the new network service bridge works there.
Additional tests will be added in the future.

While here, I moved away from using the home dir bind mount as a linter
was complaining.

BUG=b:201774934
TEST=tast run <DUT> filemanager.DrivefsGoogleDoc
TEST=tast run <DUT> filemanager.DrivefsGoogleDoc.chrome_networking

Change-Id: I678f516a8cbc6c4b53f61709c8106ff18ac6f2bd