tast-tests: Take screenshot and retry on Play Store empty error
Play Store has a bug where it doesn't show the installed apps in the
catalog view so we're going to retry the test when that bug is hit and
also take a screenshot when this happens.

Currently screenshot was only taken when all attempts have exhausted but
usually that happens when the test times out so it is too late and we
don't get the details for each failed assertion.

BUG=b:231751280
BUG=b:186274474
TEST=./fastbuild.sh -T
TEST=tast run DUT enterprise.ARCProvisioning.vm_unstable
TEST=forcefully fail the test and ensure screenshot is placed in faillog

Change-Id: Icff38cfb7a11e7dd535e3425df1a0a639a8e4987