video: Add new case and broken vector to ChromeStackDecoderVerification
This CL adds a new test case "h264_broken_files" to the MD5SUM
verification, and a first test vector to it. This test vector
has a bitstream bug and fails to verify (hw decoder vs the sw
decoder reference) in all platforms I've tried, so is best to
isolate it from the existing "h264_files_from_bugs" test case.

BUG=b:234651916
TEST=ChromeStackDecoderVerification.h264_broken_files on dedede,
trogdor, hana

Change-Id: I6888ee6199f4719e08c6dce757d099428addb06a