tast-tests: WiFi: Repeat the active scan in wifi.ChannelScanDwellTime
Probe requests from the DUT may not be captured in the pcap file
as the AP keeps sending beacon frames, which fails the test.
According to b/225073589, the loss rate is about 3%.

Repeat the active scan in wifi.ChannelScanDwellTime if the probe
request is lost. Rerunning the test gives <<1% failure rate which
is acceptable.

BUG=b:225073589, b:231669774
TEST=tast run -var=router=localhost:2210 -var=pcap=localhost:2220
localhost:2200  wifi.ChannelScanDwellTime

Change-Id: I98963de09bf2f32f8c113451de50a5fad9665b99