tast-tests: Do not launch the Play Store before apps install
Launchig the Play Store soon after provisioning can sometimes result in
failure to load native libs in Play Store and that results in different
kinds of problems e.g. apps not installing or catalog view being empty.

In this change we're also removing the opening/closing of Play Store
since it was only necessary if the store was open when the apps were
installing.

BUG=b:231751280
BUG=b:186274474
BUG=b:228355858
TEST=./fastbuild.sh -T
TEST=tast run DUT enterprise.ARCProvisioning.vm_unstable

Change-Id: I3f515550dffc4073ed51c9ba180d8921cf8f92b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3689150
Reviewed-by: Luzanne Batoon <batoon@google.com>
Reviewed-by: Alexander Hartl <alexanderhartl@google.com>
Commit-Queue: Muhammad Hasan Khan <mhasank@chromium.org>
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>