tast-tests: change Swedish to English glideTyping
Noticed that the glide typing test is flaky with Swedish in recent CQ builds: http://b/234882294. For some reason, the test results never
surface to stainless (will take a look at this).

The failure reason is similar to the previous Spanish as in https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3574183

Change Swedish to English, since the test is validating the functionality
of glide typing, no requirement for any specific language.

BUG=b:234882294
TEST=tast run inputs.VirtualKeyboardGlideTyping*

Change-Id: Ic08e8097147f056042885f319abff221ee589d9b