Add relevant tests to drivefs-cq to run on google3 cq
Having DriveFS tests run in google3 is imperative to catching
regressions as close to the source of change as possible. To avoid
having a long selector list, opt for our dormant group:drivefs-cq as the
selector for relevant tests to be ran presubmit for drivefs changes on
google3.

BUG=b:222514144
TEST=tast list '("group:drivefs-cq")'

Change-Id: Ib55106076ad04877af5594e015fc313c9e3f6e54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3686953
Reviewed-by: Travis Lane <travislane@google.com>
Tested-by: Ben Reich <benreich@chromium.org>
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Commit-Queue: Ben Reich <benreich@chromium.org>