network: Add http server to virtualnet package
This builds upon the virtualnet package and adds an http server. The
captive portal environment creates a separate network namespace
connected to the global namespace through a virtual ethernet pair. An
http server and dns server are created in the new network namespace.
When a portal recheck is called in shill, the dns will reroute the http
call to the http server, which will cause the service state to be
changed to online.

BUG=b:234866776
TEST=tast run ${DUT} network.ShillCaptivePortalHTTP, checked that
rechecking portal on veth will change the service state to online

Cq-Depend: chromium:3670620
Change-Id: I763bf185e3a471edf870972105a38f39897575fc