tast-tests: Dump bugreport for each attempt
The current implementation only writes bugreport when s.HasError is true
which is only when we use s.Fatal and that happens only on the last
attempt.

With this change, we'll dump bugreport on each failure inside the poll
as long as it is after the bugreport initialization code.

BUG=b:186274474
TEST=tast run DUT enterprise.ARCProvisioning.vm_unstable

Change-Id: I94465233d208e10ccafebd58ade6739cb28cc16d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3690165
Tested-by: Muhammad Hasan Khan <mhasank@chromium.org>
Auto-Submit: Muhammad Hasan Khan <mhasank@chromium.org>
Commit-Queue: Luzanne Batoon <batoon@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Luzanne Batoon <batoon@google.com>