network: Add routing_high_priority test
This patch adds the first routing test which verifies the routing setup
in the case that a network with higher priority shows up. The new added
routing package contains the utils that will be shared by multiple
routing tests, which will be added in the follow-up patches.

BUG=b:192436642
TEST=tast run $DUT network.Routing*

Change-Id: I7a92857088dc47cdba7e693100e2be383f97ea1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3689373
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>