tast-tests: do not verify logs for appsplatform.Dgapi2OneTime
The underlying test app is unstable and can sometimes fail to execute
an SKU purchase. This will break the test, because we look for purchase
confirmation in the logs. As we intend to test the purchase flow,
purchase confirmation is not required for the test to pass and could
be skipped.

Do not look for purchase confirmation in the logs in the test.

BUG=b:235022481
TEST=tast run dut appsplatform.Dgapi2OneTime.vm

Change-Id: Id6bc2821354143de3ee6ad315ea6d100fd5852a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3686831
Commit-Queue: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Jeevan Shikaram <jshikaram@chromium.org>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Tested-by: Alexey Shpakov <ashpakov@google.com>