network: virtualnet: Read ip addrs on ifaces
This patch adds the code to read the ip addrs on the interfaces from the
kernel. It is required because some addresses are unknown to the test
codes (e.g., IPv6 addresses acquired by SLAAC).

Also adds a util function which can be used to enter the netns
associated with the env.

BUG=b:192436642
TEST=../platform/tast/fast_build.sh -T

Change-Id: I76b16a5852a6584ff4c4c3e7b2cedd8330bea1bc