[tast-tests] Overflow shelf smoke test
This CL implements a smoke test for the overflow shelf, i.e. the shelf
that accommodates so many shelf icons that these icons cannot show at
the same time and some of them are hidden.

In detail, the overflow shelf is verified through the following steps:

Verify the bottom overflow shelf:
  1. Add enough icons so that the shelf enters overflow mode. Verify
     that the arrow button shows at the expected location.
  2. Pin additional app (such as the Settings app) through the context
     menu. Verify that the pinned app is visible and the arrow button
     shows at the expected location.

Verify the vertical shelf, i.e. the shelf that is placed at the left or
right side of the display, under the overflow mode as well. Different
from the bottom shelf, the vertical shelf is not affected by RTL.

BUG=b:200987018
TEST=shelf.OverflowShelfSmoke.*

Change-Id: I6037ab59ea42044539f5cfcf9e83d3e1328b796a