Create mouse movement simulation for CUJ tests
Simulate moving the mouse during a test by moving it in a spiral
according to the function r = theta. Once the mouse reaches the
end of the screen, the mouse resets position to the center. By
using this function, we can avoid specifying predetermined
coordinates for the mouse to move to, as well as progressively
moving the mouse faster until we hit the edge of the screen.
This gives CUJ tests an easy way to measure mouse input latency
metrics. Initially, this change adds mouse simulation while idling
in MeetCUJ.

BUG=b:234094482
TEST=Run tast run <DUT> ui.MeetCUJ.2p ui.MeetCUJ.lacros_2p, and verify
results-chart.json has values for Event.Latency.EndToEnd.Mouse.

Change-Id: I4d2e15fc091836a71d10bb1dde8c6b876a3ed1a5