camera: AUE simulation test
There were multiple incidents where a vendor forgot to remove
"free trial" logic from their closed-source library.
This test suite is to simulate the "future", ideally near the "Auto Update Expiration" time.
libfakedatetime is utilized for this purpose (cros-camera-test). Make sure that the
libfakedatetime.so is in the LD_LIBRARY_PATH directory.

BUG=b:228811937
TEST=tast run <BOARD-IP> camera.HAL3AUE

Change-Id: I80c3a9e6cc9014d3ca5cb0a9032828a0234ecf6a