familylink: add Lacros variant to policy Tast test
Add a Lacros variant to the policy Tast test. This
change extends the existing test to verify that the
policy is recognized in the browser. It also changes
the policy used to test, since safe search is on by
default for child users which would cause the
browser portion of the test to vacuously pass.

BUG=chromium:1313166
TEST=both variants pass

Change-Id: I9a757ae6c2db55cb1a836ac7cb555c549ec32766