u2fd: Change LeftClick to DoDefault
There are some coordinate resolution issues observed in Lacros, and the
new guideline is that we shouldn't rely on coordinates to click an
object. Instead, we should use `DoDefault` that triggers the default
action of an element in the future. Switch to use DoDefault in WebAuthn
procedures when clicking a button.

BUG=b:235048980
TEST=u2fd.WebauthnUsingPassword.tpm1_lacros turned from fail to pass
after this CL on my DUT
TEST=tast unit test passed

Change-Id: I0d2266f5cce71bd8d8d91ef1593700adbe1af8db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3694814
Reviewed-by: Seewai Fu <seewaifu@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>