tast-tests: fix ResizeBackupRestore and BackupRestore cleanup
ResizeBackupRestore.* tests delete a part of the image in order to
speed up the test execution. As a result, subsequent tests which
depend on these files might fail. There are two options available:
1. Remove the optimisation
2. Introduce a fixture that will automatically reset itself after test
It is not clear, if 1) or 2) is more time efficient due to different
boards performance. However, 2) is more versatile and could be reused
in other scenarious.

Introduce fixtures that reset themselves after every test. Use them
in ResizeBackupRestore.* and BackupRestore.* tests.

BUG=b:234523541
TEST=tast run dut crostini.ResizeBackupRestore.stable\
 crostini.SecureCopyPaste.copy_wayland_buster_stable

Change-Id: I5c9bf7244b02b942ca6ddd3d773396e40d5f720a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3692372
Tested-by: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: Alvin Jia <alvinjia@google.com>
Commit-Queue: Alexey Shpakov <ashpakov@google.com>
Reviewed-by: James Ye <jamesye@google.com>