lacros: Iterate and record average in DocsCUJ
Instead of interleaving the DocsCUJ benchmark in ash and lacros, this CL
now uses one variation and loop through the test using the same type of
browser. This resembles a more realistic scenario than what was
implemented before.

Besides this CL helps to address the spurious data of DocsCUJ seen in
the Crosbolt dashboard by adding iteration runs in order to record the
average.

BUG=chromium:1328923
TEST=tast run device lacros.DocsCUJ*

Change-Id: Icaa2dc6ec7aed95905b63334f858c9c5e143ab56