tast-tests: Checks more command line for browser-process.
Processes with (one of) --help, --h, --product-version, --version,
--credits are not the browser process, so exclude them from the candidate
of browser-process.

Actually, e.g. crash_reporter runs ash-chrome with --product-version
to get its version info. Such a process was sometimes found as a candidate
of browser-process, and so test fails to identify the browser-process,
which must be unique on the system.

BUG=b:231683759
TEST=Tryjob. fast_build.sh -T.

Change-Id: I69e999153075b17b63cdf311a06e9b9f5377cfcf