tast-tests: Wait for bounds conditions in arc.SplitView
This is one of the attempts to stabilize the arc.SplitView test.
On low-end devices and/or when the CPU load is pretty high, the
window bounds change is delayed for a while so sometimes the
one-time bounds check fails which causes the flakiness.
So this CL changes to keep polling until the window bounds
condition gets satisfied.

BUG=b:233559806
TEST=tast run arc.SplitView*
TEST=tast unit tests passed.

Change-Id: Ia4dcf78c6eccc9bc52fdf35aae5c5f6a6fd3528a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3695648
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Toshiki Kikuchi <toshikikikuchi@chromium.org>
Tested-by: Toshiki Kikuchi <toshikikikuchi@chromium.org>