tast-tests: refactor hwsec.PrepareCrossVersionLoginData
Currently, we only test the migration from specific version to current
version. It would be very useful to test the migration on the same
device with current version, so that we could see if the failure is due
to the preparation, verification or the actual migration bug. To do so,
we need to refactor the preparation and the verification test.

This CL is for refactoring the preparation part.

BUG=b:228279581
TEST=tast run <ip> hwsec.PrepareCrossVersionLoginData

Change-Id: I575eb3545d0c5e9210963a750de865aee6065860