Launch notification test app by launcher and grant notification permission
Notification.go test is failing on T because lacking of permission to
change android notification. On T it is asking for permission to display
notification.

To fix the issue, launch the test app from the launcher, so that the
GrantPermissionsActivity of Notification can show. Then, click the
"ALLOW" button to grant notification permission to the test app. If it
fails to find the permission dialog, log the error but continue the test
so that with the versions without the permission dialog, the test can
still function normally.

BUG=b:234760125
TEST=tast run arc.Notification.vm

Change-Id: I0aebd3cfbfdb88c5d1f6fa5c0502010b490c9ae1