Create key press simulation for CUJ tests
Simulate pressing a single key repeatedly for a given amount of time.
This allows for easy key input in tests that can use a single button
to perform actions, such as scrolling in GoogleSheetsCUJ,
GoogleSlidesCUJ, QuickCheckCUJ, and MeetMultitaskingCUJ.

BUG=b:234094482
TEST=Run each of the affected tests, and ensure that
Event.Latency.EndToEnd.KeyPress is present in results-chart.json.
Additionally, ensure that visually the tests look correct (the delays
between key presses are reasonable and appropriate).

Change-Id: Ie921f6a53fe7a239e05fc7d7c79a296a04301534