tast-tests: refactor hwsec.CrossVersionLogin
Currently, we only test the migration from specific version to current
version. It would be very useful to test the migration on the same
device with current version, so that we could see if the failure is due
to the preparation, verification or the actual migration bug. To do so,
we need to refactor the preparation and the verification test.

This CL is for refactoring the verification part.

BUG=b:228279581
TEST=tast run <ip> hwsec.CrossVersionLogin*

Change-Id: I01d4e19fb9ee1231ba27791aaf86fc17b877c8cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3697861
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Ching-Kang Yen <chingkang@chromium.org>
Tested-by: Ching-Kang Yen <chingkang@chromium.org>
Reviewed-by: Leo Lai <cylai@google.com>