tast tests: Fix the blocked apps test
The test has been checking the screen for
install button not to be present. Instead
this should check if install (greyed out)
is present so that play store loading delay
isn't counted as Pass where install button
is not present.

BUG=b:237023275
TEST=tast run <DUT> arc.UnicornBlockedApps.vm (Verify that test fails
when playstore fails to load).

Change-Id: Iad4968d2208b7ab25e188ecac351cf2b8834e004
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3723197
Reviewed-by: David Welling <davidwelling@google.com>
Commit-Queue: Renuga Nanjappan <rnanjappan@chromium.org>
Tested-by: Renuga Nanjappan <rnanjappan@chromium.org>