tast-tests: Add hwsec.CrossVersionLogin.current
Currently, we only test the migration from specific version to current
version. It would be very useful to test the migration on the same
device with current version, so that we could see if the failure is due
to the preparation, verification or the actual migration bug. To do so,
we need to refactor the preparation and the verification test.

This CL is for the implementation of the verifying migration with the
current version.

BUG=b:228279581
TEST=tast run <ip> hwsec.CrossVersionLogin*

Change-Id: I253c27738c33c32b806f0db748967132602f4e29