tast-tests: Set bot layout for MeetCUJ 2p and 4p tests
We need to set the correct layout, otherwise the bot will not request a
high resolution outbound video. The 2p test is still limited to 360p.
In order to get 720p I need to manually click the settings and set the
video mode to High.

BUG=b:233627706
TEST=tast run $IP ui.MeetCUJ.2p and 4p
Verify outbound resolution jumps from 180p to 360p.

Change-Id: I6c52432ab882b79c76a157744e14ef9ac122e554