[tast-test] Fix the ScreenBrightnessPercent test
In this tast test, it should click on the unified system tray instead
of the  status area. The status area carries a lot of trays, such as
the  dictation tray, keyboard tray and etc. This test passed before
because the middle of the area of all the trays happened to be on the
unified system tray.

It failed in the associated bug because after adding the date tray, the
middle of the status area widget is not right on the unified system tray
any more.

BUG=chromium:1335176
TEST=policy.ScreenBrightnessPercent

Change-Id: I9c775315baf3b86726d8f5da2b7515d864ec33a5