lacros: Change to use lacros-chrome-additional-test-args instead
Originally, Tast uses `lacros-chrome-additional-args` to add lacros
specific arguments in command line. However, since now we might add it
in ChromeCommandBuilder, the one Tast added might break the one added by
ChromeCommandBuilder.

As a result, this CL changes to use `lacros-chrome-additional-test-args`
instead in Tast so that they can be merged instead of replacing each
others on the Chrome side.

BUG=b:233692321
TEST=Manually

Change-Id: Iad971c05537e25f62fdbc3cf03cf25d3f17e209a