Add router reinit and call it between wifi tests
This change adds a new ReinitTestState function to the base
Router interface that all router controllers must implement,
and adds a basic no-op implementation of it to the ax, openwrt,
and legacy router implementations. ReinitTestState is intended
to allow for a router controller implementation to reinitialize
their state between tests to make them more stable. This new
function is called once for every router controller in the
test fixture, including the pcap router, in the Reinit function
of the wifi test fixture.

This also updates the wifi test fixtures to call the fixture
Reinit function during PreTest instead of during Reset to
ensure that the router state is as clean as possible prior
to the test call. In addition, it is now expected that the
Reinit function will be called between subtest runs for tests
that make many router setting changes, and wifi.SimpleConnect
has been updated to do so.

BUG=b:211517577
TEST=Successfully ran wifi.SimpleConnect.80211a and confirmed
that ReinitTestState was called once for each router controller
between test cases as expected by reviewing the test logs.

Change-Id: Iab3ea379f963d8cd46a396ae8e879d32ddf3468e