tast: dnsproxy: Refactor and cleanup
This patch refactors the support code for testing DNS by first
separating out dns-proxy specifics from general logic. It also
introduces a new design for working safely with iptables,
automating the removal of the rules. This can again be generalized
in a follow-up patch to be used anywhere, not just for DNS.

BUG=b:230686377
TEST=tast run <DUT> network.DNSProxy*

Change-Id: I7d611f3b43193744ac6297c3c6b64da6a3ede0a3