fp_tpm_seed: Don't check if biod is running
Until now, firmware.FpTpmSeed was checking biod service status to get
information if bio_crypto_init provided TPM seed to FPMCU. It could
happen that biod was stopped by some other test, which caused this test
to timeout on waiting for biod.

On the other side it's not necessary to check if biod is running to
determine if bio_fw_updater provided TPM seed. The bio_crypto_init
blocks the boot process in boot-services (boot-services won't be marked
as 'started' until bio_fw_updater completes). The system-services can
start only when boot-services are started. The failsafe starts when
system-services are starting and finally openssh-server starts when
failsafe is starting. As a result, if the test has SSH connection to
DUT then we are sure that bio_fw_updater completed.

BUG=b:233276848
TEST=tast run <IP> firmware.FpTpmSeed

Change-Id: I1c456f86b058ce82b6c3e870d3012e5f90fa263c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3702364
Tested-by: Patryk Duda <pdk@semihalf.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>