Extend force-sync during onboarding to Smart Lock tests
Try to improve onboarding stability for Smart Lock tests by applying the
force sync step that we use in the Phone Hub ones. If we're stuck in the
state where the "Verify" button is shown under Connected Devices in OS
Settings for 5 minutes, we'll force a DeviceSync through the
chrome://proximity-auth debug page.

BUG=b:230401333
TEST=tast run -var=crossdevice.PhoneIP="192.168.86.23" localhost:2223
'("group:cross-device")'

Change-Id: Ia9975ddf7770491cac00f29fd9b76214ff80ca30