fp_tpm_seed: Check if /run/bio_crypto_init/seed does not exist
File that contains TPM seed (/run/bio_crypto_init/seed) is created in
tmpfs by cryptohome. It is used by bio_fw_updater to initialize FPMCU
with TPM seed. When the file is no longer needed it has to be removed.

This patch adds a check for the presence of a TPM seed file. The test
will fail if the file exists.

BUG=b:233276848
TEST=Run 'tast run <IP> firmware.FpTpmSeed' make sure test passes.
TEST=Create TPM seed file using 'touch /run/bio_crypto_init/seed'.
     Run 'tast run <IP> firmware.FpTpmSeed' make sure test fails.

Change-Id: I0c69c40402f46ad6680659ad7d8d7d3fc00fa514
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3706820
Reviewed-by: David Welling <davidwelling@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Patryk Duda <patrykd@google.com>
Reviewed-by: Bobby Casey <bobbycasey@google.com>