Projector: Fix creation flow Tast test failures
This CL does the following:
* Removes clicking the maximize button because it doesn't exist on
tablet devices.
* Adds microphone hardware dep to the projector.CreationFlow Tast test
to ensure that the new screencast button isn't disabled due to a lack
of a microphone on the device.
* Adds a new test projector.NewScreencastButtonCondition that checks
if the device has SODA installed and microphone available.
If so, then the test verifies the new screencast button is enabled.
Otherwise, the test checks that the button is disabled with the
appropriate error tooltip, and prematurely exits the test.
* Refactors some helper functions into projector.go and a11y.go
utility files.

BUG=b:236155644
TEST=./repeat-test.sh -r 10 -- -buildbundle=cros localhost:9222 projector.*

Change-Id: I5d941cc73eae2f11ca9c6b1cd225eda478ec60ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3707905
Commit-Queue: Toby Huang <tobyhuang@chromium.org>
Tested-by: Toby Huang <tobyhuang@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>