video: split 4K videos out of h264_files_from_bugs
ChromeStackDecoderVerification.h264_files_from_bugs has one sequence
that is 4K in size. This CL splits that into another test case so
that it can be correctly gated by capabilities.

At least on asurada, which doesn't support 4K, that sequence is
decoded as 1080p, causing grief and crashes due to accessing
memory out of bounds (there's another WIP to prevent that in
production).

BUG=b:230377834
TEST=ChromeStackDecoderVerification.h264_4k_files_from_bugs skips
on e.g. asurada

Change-Id: Ic5cc36ce97ba498888070f27e69d4a033f029eb2