tast: dnsproxy: Add tast for custom nameservers
Adds a tast to verify that when a plaintext query is made to a specific
nameserver, i.e. one that is explicitly provided and different from the
nameserver configured for the network, that it will be used directly and
dns-proxy will by bypassed.

Some additional refactoring is necessary to share functions between
tasts.

BUG=b:230686377
TEST=tast run <DUT> network.DNSProxyCustomNameserver

Change-Id: Ifd4bfe8d931bc9cfeefbf003ee24a75e79249671
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3707028
Commit-Queue: Garrick Evans <garrick@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: Garrick Evans <garrick@chromium.org>