tast-tests: wait for focus after clicking a node
Instead of checking focused status just after clicking a node, this CL
makes the test wait for the focused state updated so that we can
eliminate the possible causes of the test flakiness.

Note that this is not the root cause of the test flakiness
reported in b:236179466.

BUG=b:236179466
TEST=tast run betty arc.StandardizedKeyboardTyping* arc.StandardizedKeyboardCopyPaste*

Change-Id: I5bd263ba8bc65d0d390cd45cf42189765bc81076
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3709383
Tested-by: Hiroki Sato <hirokisato@chromium.org>
Auto-Submit: Hiroki Sato <hirokisato@chromium.org>
Reviewed-by: David Welling <davidwelling@google.com>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Hiroki Sato <hirokisato@chromium.org>