tast-tests: wait for focus after clicking a node
Instead of checking focused status just after clicking a node, this CL
makes the test wait for the focused state updated so that we can
eliminate the possible causes of the test flakiness.

Note that it's not clear if this is the root cause of the test flakiness
reported in b:236179466.

BUG=b:236179466
TEST=tast run betty arc.StandardizedKeyboardTyping* arc.StandardizedKeyboardCopyPaste*

Change-Id: I5bd263ba8bc65d0d390cd45cf42189765bc81076