hps: adjust test expectations based on whether HPS is present
On *-generic, we expect hpsd to be running with a fake peripheral.
On brya, hpsd will only be running if the HPS peripheral is physically
present. HPS is only included in a certain SKUs of certain boards.

Rather than trying to skip the test based on particular boards or SKUs,
let the test always run and adjust its expectations based on whether
'cros_config /hps has-hps' indicates the HPS periphal is present or not.
This should get the test passing on all *-generic and brya boards.

BUG=b:214844812
TEST=tast run $BRYADUT hps.DBus hps.Running
TEST=tast run $BETTYVM hps.DBus hps.Running

Change-Id: Ib6ee8a921467449a8b7af3c08fe8ab4092e2d0ed