Remove disabling OobeConsolidatedConsent from ui.ChromeVariationsSmoke
OobeConsolidatedConsent moved the EULA acceptance to be marked
done after the first user sign in, this caused
ui.ChromeVariationsSmoke test to fail.

As a temporary fix, OobeConsolidatedConsent feature was disabled
in the test. Now that crrev.com/c/3641765 is landed,
ui.ChromeVariationsSmoke is now passing with the feature both
disabled or enabled.

BUG=b:233803343
TEST=tast run -build=true <ip> ui.ChromeVariationsSmoke

Change-Id: I5d0a270f6ebcd4cd5ea787f21bba14069f3e2a28