testexec: Handle wrapped ExitError
crrev.com/c/3157892 wrapped the result of cmd.Wait() in some cases.
This means that casting it to exec.ExitError in testexec.GetWaitStatus()
fails.  We can't use errors.Unwrap() because other cases still return
the bare error from Wait.  Instead, switch to errors.As() so that both
cases are handled.

BUG=b:175631323
TEST=tast run printer.RejectDocumentFormat

Change-Id: I7fd9d2dd42d23f0fd18ea8965e4e4cbfb038f209