video: Use graphicsNoChrome fixture
The graphicsNoChrome provides a way to keep tests from starting/stopping
Chrome for each test case. It also provides dmesg monitoring with the
gpuWatchDog parent fixture. These are both useful for test cases that
need to run without Chrome.

This CL uses the graphicsNoChrome fixture for tests that previously
stopped the UI and restarted it after each test case.

BUG=b:235879582
TEST=video.* on Trogdor

Change-Id: I1571689b163918a8edd925feb25d0f7bad44f6d2