Add keyboard input to VideoCUJ
Augment VideoCUJ to press and hold the up and down arrow with the video
playing, thus changing the volume, to capture input latency with regards
to holding down a specific key. Additionally, pause and play the video
with the spacebar to capture general key-press input latency.

BUG=b:236030517
TEST=Run ui.VideoCUJ.clamshell and ui.VideoCUJ.lacros and verify
that each results-chart.json has values for
Event.Latency.EndToEnd.KeyPress.

Change-Id: I915338b38b2c4c576c9c64a287c6c6f67a2338b4