tast-tests: Add network verification during OOBE
Update rollback service to check the preservation of the networks
during OOBE as well.

Create verifyNetworks function that receives the corresponding
connection to the api and verifies the networks. Call the method
during OOBE and after login to check the networks set are the
expected ones in both cases.

BUG=b:229103041

TEST=tast run DUT policy.EnterpriseRollbackInPlace
tast run -var=updateutil.tlwAddress="localhost:7151" DUT
autoupdate.EnterpriseRollbackPreviousVersion.rollback_1_version

Change-Id: I25916bb56284e645eb0d1fea48f67ed348f2dbc0