ui: Fix failing tests ui.ChromePIPEnergyAndPower.*_lacros
The tests are failing in the lab, but they pass when I run them myself,
even if I make a careful choice of device and ChromeOS version. I think
the failures must be caused by improper cleanup in a test that runs
sometime before, but I have no way of knowing what test is guilty of
such improper cleanup. This CL gives ui.ChromePIPEnergyAndPower.*_lacros
their own separate copy of the lacros fixture.

BUG=b:237109034
TEST=tast run <DUT> ui.ChromePIPEnergyAndPower.*_lacros

Change-Id: I2d13398e738de3e015781b483e85f46ea7a838d7