power: Switch battery discharge over to chargeoverride command
By using the chargeoverride command we get udev notifications
and the the charger status is correctly reflected in sysfs.

Without this patch:
    $ tail /sys/class/power_supply/CROS_USBPD_CHARGER0/online
    /sys/class/power_supply/CROS_USBPD_CHARGER0/status
    ==> /sys/class/power_supply/CROS_USBPD_CHARGER0/online <==
    1

    ==> /sys/class/power_supply/CROS_USBPD_CHARGER0/status <==
    Charging

With this patch:
    $ tail /sys/class/power_supply/CROS_USBPD_CHARGER0/online
    /sys/class/power_supply/CROS_USBPD_CHARGER0/status
    ==> /sys/class/power_supply/CROS_USBPD_CHARGER0/online <==
    1

    ==> /sys/class/power_supply/CROS_USBPD_CHARGER0/status <==
    Not charging

This also involved reworking the way the fwupd tests worked. It was
also trying to manage the battery. I refactored it to use a cleanup
function instead. I'm not honestly sure why we are calling stressapptest
as part of switching to DC.

BUG=b:233778746
TEST=tast test $IP ui.MeetCUJ.4p

Change-Id: I6a93ab3c9cbbdadcfea83252c4dd09174a9d88e6