network: Add the tcpdump utility to the local test environment
This modification adds the tcpdump utility in the local test
environment. It extracts the logic related to the tcpdump
command from the cpature.go file so that both local and remote
tests can use the tcpdump utility.

BUG=b:237283936
TEST=Run wifi_matfunc to check that all tests pass and
tast run ${DUT} network.ShillCaptivePortalDNS to check
tcpdump utility can be used in local test environment and
run fast_build.sh to rebuild the tast executable.

Change-Id: I21a50338b9a59bbf0f8611da12000cbab4d8ca68