wifi: Add auth type to Passpoint credentials
This patch adds the auth type of a Passpoint credentials. This is
necessary to test provisoning using EAP-TLS. Alongside that, this
patch also moves the credentials to credential.go in order to
further extend the functionality to create Passpoint ARC config
in a follow-up patch.

BUG=b:229700842
TEST=tast run <DUT> wifi.PasspointRoaming.*
TEST=tast run <DUT> wifi.PasspointSelection.*

Change-Id: I3e749ad332f142c70249c056943b0585f07abe1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3727299
Reviewed-by: Avery Musbach <amusbach@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Damien Dejean <damiendejean@google.com>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>