lockscreen: Assert the correct attempts to lockout PIN
This test checks whether the PIN input is hidden within the maxAttempts,
which is set to 10. This limitation is too loose, we should assert that
PIN input isn't hidden within 5 attempts, and is hidden exactly aftet
the 5th incorrect attempt, so we could catch bugs like the attached bug
in the future.

Due to the mentioned bug, PIN's auth_locked will only be set after N+1
incorrect attempts, so set the lockoutAttempt to 6 temporarily so that
the test passes.

BUG=b:234715681
TEST=tast run $DUT lockscreen.PinUnlockFail

Change-Id: I17ff7af927f2ff220fa14f2aeecaebc21aa5a004