hps: use hwdep.HPS() to skip on DUTs without HPS
On *-generic, we expect hpsd to be running with a fake peripheral.
On brya, hpsd will only be running if the HPS peripheral is physically
present. HPS is only included in a certain SKUs of certain boards.

Split hps.DBus into two parametrized tests.

The 'hps.DBus' test (unnamed parameter) is the "real" version which uses
hwdep.HPS() to ensure it is only run on DUTs where HPS is physically
present.

The 'hps.DBus.fake' test runs on VMs, where hpsd is configured to use
a fake peripheral. This is important for giving us test coverage of hpsd
in the short term, while we don't have any lab DUTs with HPS.

BUG=b:214844812
TEST=tast run $BRYADUT hps.DBus hps.DBus.fake
TEST=tast run $BETTYVM hps.DBus hps.DBus.fake

Cq-Depend: chromium:3721660
Change-Id: Ib6ee8a921467449a8b7af3c08fe8ab4092e2d0ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3714530
Tested-by: Dan Callaghan <dcallagh@chromium.org>
Commit-Queue: Dan Callaghan <dcallagh@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>