drivefs: Add a test that uploads a blob
This adds filemanager.DrivefsBlobUpload which writes a text file to
DriveFS and ensures that the file is uploaded to Drive Web. It then
compares the MD5 sum of the uploaded data with what is available
locally.

I've added some other support utilities to the drivefs package which
were used in writing this test, or will be used in future tests.

BUG=b:201774934
TEST=tast run <DUT> filemanager.DrivefsBlobUpload
TEST=./fast_build.sh -t chromiumos/tast/local/drivefs

Change-Id: Ie6a96c8e418123de55f4f737eebcc6639201a1b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3727460
Tested-by: Travis Lane <travislane@google.com>
Reviewed-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>