tast-tests: Fix the quick-settings button class name
In this tast test, it should click on the unified system tray instead
of the status area. The status area carries a lot of trays, such as
the  dictation tray, keyboard tray and etc.

It failed after calendar view is enabled because after adding the date
tray, the middle of the status area widget is not right on the unified
system tray any more.

By fixing the class name it will fix the error:
`quick-settings-0: Failed to run the test scenario:
failed to proceed the test scenario` in the fail report in the
associated bug. But there's still a "[ FAIL ] Performance expectation
missed: Apps.ClamshellLauncher.AnimationSmoothness...". So I didn't
enable the test. Will reassign this bug to the related owner after this
cl is submitted.

Also changed the deprecated `ClassName` to `HasClass` in this file

BUG=chromium:1335176
TEST=tast run <DUT> ui.DesktopControl
Change-Id: I796f1c204ed28eb5cf13a75de051a5e5ec5f7f67
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/3731063
Tested-by: Jiaming Cheng <jiamingc@chromium.org>
Commit-Queue: Jiaming Cheng <jiamingc@chromium.org>
Reviewed-by: Kamil Szarek <kamilszarek@google.com>
Reviewed-by: Avery Musbach <amusbach@chromium.org>