tast-tests: Cleanup files older than 1 hour on Drivefs tests
DriveFS tast tests re-use a pool of accounts that have drive enabled. In
the case that the cleanups don't run properly for each individual
account or a cleanup isn't added, we want to ensure the accounts don't
exceed their space. This adds a post fixture teardown task to cleanup
all files older than 1hour.

Note the file removal can take ~1s per file to remove. The current
timeout for the TearDown is 1 minute, so if there are more than 60 files
this may exceed the timeout. This is ok as this cleanup will run on
every teardown and thus will progressively clean up the account.

BUG=chromium:1299712
TEST=tast run filemanager.DrivefsGoogleDoc

Change-Id: I1ed0db9a5686432d99b97bbb6c31ad5adbdf6f1a