tast-tests: Do not enroll in EnterpriseRollbackInPlace
Remove enrollment from test. The test do not check any feature
dependent from enrollment and a big number of failures are caused
by this dependency. Refer to the corresponding bug for further
details.

The test is moved to the mainline group and added the informational
tag to test its stability before adding it to the CQ.

BUG=b:236359370

TEST=tast -verbose run DUT policy.EnterpriseRollbackInPlace

Change-Id: Ia7404063600c60b664f29a31b37102a7d404e3e5