drivefs: Add a test that downloads while offline
This adds filemanager.DrivefsBlobDownloadOffline which creates a file
with the Drive API, waits for the file to appear in DriveFS, disables
the network, and then attempts to read the file.

This must fail in a timely manner, as blocked reads in fuse tend to lead
to deadlocks and system instability.

BUG=b:201774934
TEST=tast run <DUT> filemanager.DrivefsBlobDownloadOffline

Change-Id: I3c3c77c5ed0b922297ebed17976dbdbaa17970b2