tast-tests: Migrate all crostini.* tests from Downloads bind mount
The /home/chronos/user/Downloads bind-mount is being migrated directly
to /home/chronos/u-HASH/MyFiles/Downloads and thus will not be
available. A presubmit has been added to discourage further usage of it,
however, we need to migrate all existing usages off. This does it for
all crostini.* tests, namely:

  - crostini.ShareDownloadsAddFiles*
  - crostini.CopyFilesToLinuxFiles*
  - crostini.DragDrop*
  - crostini.ShareFolderZipFile*
  - crostini.ShareFolders*

BUG=b:233132576
TEST=tast run all tests listed above
TEST=./fast_build.sh -T

Change-Id: I409adc569fcc634314ea2cf2317ff15c8fcfab50