tast-tests: Migrate apps.* test from Downloads bind mount
The /home/chronos/user/Downloads bind-mount is being migrated directly
to /home/chronos/u-HASH/MyFiles/Downloads and thus will not be
available. A presubmit has been added to discourage further usage of it,
however, we need to migrate all existing usages off. This does it for
all apps.* tests, namely:

  - apps.ContentPreview*
  - apps.LaunchGallery*
  - apps.Sharesheet*

BUG=b:233132576
TEST=tast run all tests listed above
TEST=./fast_build.sh -T

Change-Id: Iffbb4313ebe7a44f5b3c0710f8a824dcd950d02b