tast-tests: Migrate nearby.* test from Downloads bind mount
The /home/chronos/user/Downloads bind-mount is being migrated directly
to /home/chronos/u-HASH/MyFiles/Downloads and thus will not be
available. A presubmit has been added to discourage further usage of it,
however, we need to migrate all existing usages off. This does it for
all nearby.* tests, namely:

  - nearbyshare.PhoneToCrosHighVis*
  - nearbyshare.PhoneToCrosBackgroundScanning*
  - nearbyshare.PhoneToCrosInContacts*

This also updates the NearbyShareService as well.

BUG=b:233132576
TEST=tast run all tests listed above
TEST=./fast_build.sh -T

Change-Id: Iea67f0696dce78d23317b74aa41639b1e402e4b0