tast-tests: Update group of tests to not use Downloads and user mount
The usage of the /home/chronos/user/Downloads bind mount is being
deprecated. This moves a chunk of tests away from using the
filesapp.DownloadPath and filesapp.MyFilesPath and to directly
retrieving the mount via cryptohome. The tests affected in this patch
are:

  - policy.DownloadRestrictions*
  - enterpriseconnectors.TestDownload*
  - enterpriseconnectors.TestFileAttached*
  - phonehub.DownloadRecentPhoto*
  - wmp.FullScreenshot*
  - wmp.RecordFullScreen*
  - wmp.RecordPartialScreen*

This also updates some associated helper methods to plumb through the
downloads path once retrieved. Specifically:

 - WaitForDMTokenRegistered
 - HasScreenRecord

BUG=b:233132576
TEST=tast run all tests listed above
TEST=./fast_build.sh -T

Change-Id: I3cd562fa8e40b3c03af8d17e31e339b7beb48fe7